fix(ci): test_sdist shall not test wheels installation (#65)
There's a race condition in the jobs meaning `dist` will be populated by the the sdist and any number of wheels. If there's more than one wheel, sdist tests will fail.
We do not require tests in test_distribution.py which are meant for local builds.
Rather than excluding these tests, simply do not provide a `dist` folder to look at. It keeps the workflow simpler and if anyone tries to reproduce locally they will run these tests which are meant to be run locally.